move config file(s)

+++ twitch/commands/test.js
@@ -1,22 +1,25 @@
 exports.run = async (client, channel, userstate, command, args, content) => {
-	client.log("warn", "RESETTING REPOSITORY TO ORIGIN/MASTER");
 	
-	const exec = require("child_process").exec;
-	
-	exec("git fetch", (err, stdout, stderr) => {
-		process.stdout.write(stdout)
-	});
-	setTimeout(function() {
-		exec("git reset --hard origin/master", (err, stdout, stderr) => {
+	if (client.config.hosted) {
+		client.log("warn", "RESETTING REPOSITORY TO ORIGIN/MASTER");
+		
+		const exec = require("child_process").exec;
+		
+		exec("git fetch", (err, stdout, stderr) => {
 			process.stdout.write(stdout)
 		});
-		
 		setTimeout(function() {
-			exec("refresh", (err, stdout, stderr) => {
+			exec("git reset --hard origin/master", (err, stdout, stderr) => {
 				process.stdout.write(stdout)
 			});
-		}, 10);
-	}, 700);
+			
+			setTimeout(function() {
+				exec("refresh", (err, stdout, stderr) => {
+					process.stdout.write(stdout)
+				});
+			}, 10);
+		}, 700);
+	}
 	
 	client.twitch.say(channel, "test successful");
 	return;
@@ -28,7 +31,7 @@ exports.config = {
 };
 
 exports.condition = (client, channel, userstate, command, args, content) => {
-	if (command === "test" && client.config.admins.includes(userstate.username)) {
+	if (command === "test") {
 		return true;
 	}
 	return false;
